compact embedding clustering procedure into a single source file
